requirements.txt
pandas
numpy
scikit-learn
xgboost
streamlit
plotly
pyarrow
joblib
pyyaml
